move ErrorMessageSelfTest
